fix: explicitly cast label column to Int64 in multiclass
- Add dtype=pl.Int64 to pl.Series creation for label column
- Fixes type Float64 is incompatible with expected type Int64 error
- Ensures label column is consistently Int64 type

+++ loop/sfm/lightgbm/utils/tradeline_multiclass.py
@@ -338,7 +338,7 @@ def create_multiclass_labels(df: pl.DataFrame,
                 labels[idx] = 2
                 short_count += 1
 
-    df = df.with_columns(pl.Series('label', labels))
+    df = df.with_columns(pl.Series('label', labels, dtype=pl.Int64))
 
     return df
